Fixed receiving arguments on services
Seems that this particular piece of code incorrectly assumes args will be an array of arguments,
while the C++ code that translates DBus Messages into V8 objects returns an array
only if there are more than one arguments. When there is only one argument, Array.prototype.slice.call
is incorrectly called on that argument.
The result is chaotic: numbers are not received at all and strings are received one character per argument.