encoder: fix arrays of items of the same type
As described in https://github.com/Shouqun/node-dbus/issues/120
A javascript array is always handled as an array of variant.

This steps through an array to confirm that all elements are of the same type,
then produces a dbus structure accordingly, now supporting:

* array of boolean
* array of byte
* array of uint32
* array of int32
* array of double
* array of string

and defaulting to array of variant otherwise.