Add basic tests against dbus
Based on the .travis.yml, it looks like this project supports back to
Node 0.12.

While doing the work for pull request #150, I used `Object.keys`.
However, I was unsure whether this was a supported call in Node 0.12.
Because the Travis-CI build only tests whether the C++ code builds
propertly, and doesn't test any of the JS code, the results of the
Travis-CI build doesn't prevent code that would break in Node 0.12 from
getting committed.

So, this commit adds some basic tests (cribbed mostly from the examples
folder) that run a service and a client against an actual dbus daemon to
test that the code can do basic things like call methods, get and set
properties, and receive signals.

The way this is done is that, each test file (`test/*.test.js`) spawns a
DBus Service (`test/service.js`) and then executes client commands
against that service.