Add bus.disconnect and service.disconnect
Allow applications to disconnect their connections to dbus after they
are done with them.

Fixes #133

Changes required to make this happen:

* Don't share bus objects: remove the code in lib/dbus.js that stores
created Bus objects and returns them when a new bus is requested.
Instead, create a new Bus object whenever one is requested. This allows
each Bus object to disconnect itself without worrying that other objects
are relying on it being connected. Note that at the C++ level, there is
connection sharing going on because of the use of `dbus_bus_get`, but
that isn't a problem because the dbus library is managing the sharing,
and calling `Bus.prototype.disconnect` will correctly notify the
underlying C++ that that particulary Bus object is no longer using its
connection.

* Remove the `dbus_connection_close` call. According to the
documentation, this method should not be used when a connection is
acquired using `dbus_bus_get`. Removing this prevents the error message
"Applications must not close shared connections - see
dbus_connection_close() docs. This is a bug in the application." noted
in #133. See [dbus_bus_get][1] and [dbus_connection_close][2].

[1]: https://dbus.freedesktop.org/doc/api/html/group__DBusBus.html#ga77ba5250adb84620f16007e1b023cf26
[2]: https://dbus.freedesktop.org/doc/api/html/group__DBusConnection.html#ga2522ac5075dfe0a1535471f6e045e1ee