Throw or return errors on disconnected connection
When an operation is attempted on a closed connection, throw a
Node-level error instead of allowing the C++ to surface an error. In
order to do this, we keep track of whether a bus is still connected or
whether it has already been disconnected.

serviceInterface.emitSignal does not have a callback and will throw an
error if the connection is already disconnected. All other methods that
try to use a disconnected connection will return the error in the
callback.