Allow name and message in errors
Set the name of the error by setting the error's dbusName property.

E.g.

    var error = new Error('This is the message');
    error.dbusName = 'test.dbus.TestService.ErrorCode';
    callback(error);

If error.dbusName is not set, it defaults to DBUS_ERROR_FAILED[1]

This matches closer with what people expect the error message to be, and
prevents the following error when a `new Error('Human readable
message')` is returned.

```
process 16776: arguments to dbus_message_new_error() were incorrect,
assertion "_dbus_check_is_valid_error_name (error_name)" failed in file
../../dbus/dbus-message.c line 1480.

This is normally a bug in some application using the D-Bus library.
```

[1]: https://dbus.freedesktop.org/doc/api/html/group__DBusProtocol.html#gabb62fd6340d0787fbd56ff8dd2f326c7